+++ src/config/index.ts
@@ -1,7 +1,7 @@
 export * from './APIRoute';
 export * from './frontendRoutes';
 export * from './majors';
-export * from './validationError';
+export * from '../validationError';
 export * from './authToken';
 export * from './constants';
 export * from './degrees';
